{VRS}{Experimental} Add a new sample to mutate VRS frame when performing a copy #13
Summary:
Demonstrate how to customize a FilterCopier to mutate images in a VRS file
We show how to decompress, manipulate the image buffer and compress it back and send the new compressed image to a new VRS file.
We are adding a class ImageMuterFunctorFilterCopier that allow to take any user Functor to modify the PixelFrame.
Many image mutation FUNCTORS are demonstrated:
- VerticalImageFlipMutator -> To vertically flip VRS images
- NullifyModuloTwoTimestamp -> To convert to black any VRS image where the timestamp is modulo2 compliant (results in converting to black some image segments in the VRS copy)
- VRSExportLoader -> To reload files exported by vrs export-image (i.e if you want blur objects in images) -> the logic to load the image and replace the pixels in the PixelFrame is left as an exercise for the user.

Reviewed By: kiransom, gustavomart

Differential Revision: D43780286

fbshipit-source-id: b1e055358cf3f9a4233920d3d0cc0d97f7e3ed08